pyproject.toml
[tool.poetry]
name = "sepana-chatbot"
version = "0.1.0"
description = ""
authors = ["Jason Bryant"]
readme = "README.md"
packages = [{include = "sepana_chatbot"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
channels = "^4.0.0"
daphne = "^4.0.0"
channels-redis = "^4.1.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
pylint = "^2.17.5"
black = "^23.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
